backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
requests==2.31.0
google-generativeai==0.3.2
python-dotenv==1.0.0
pydantic==2.5.0
motor==3.3.2
websockets==12.0
aiofiles==23.2.1